+++ src/core/time.h
@@ -7,25 +7,25 @@
 
 namespace barely {
 
-[[nodiscard]] constexpr double FramesToSeconds(double sample_rate, int64_t frames) noexcept {
-  assert(sample_rate > 0.0);
-  return static_cast<double>(frames) / sample_rate;
+[[nodiscard]] constexpr double FramesToSeconds(float sample_rate, int64_t frames) noexcept {
+  assert(sample_rate > 0.0f);
+  return static_cast<double>(frames) / static_cast<double>(sample_rate);
 }
 
-[[nodiscard]] constexpr int64_t SecondsToFrames(double sample_rate, double seconds) noexcept {
-  return static_cast<int64_t>(seconds * sample_rate);
+[[nodiscard]] constexpr int64_t SecondsToFrames(float sample_rate, double seconds) noexcept {
+  return static_cast<int64_t>(seconds * static_cast<double>(sample_rate));
 }
 
 [[nodiscard]] constexpr double Quantize(double position, int32_t subdivision,
-                                        double amount) noexcept {
+                                        float amount) noexcept {
   assert(subdivision > 0);
-  assert(amount >= 0.0 && amount <= 1.0);
+  assert(amount >= 0.0f && amount <= 1.0f);
   return std::lerp(
       position,
       static_cast<double>(static_cast<int64_t>(static_cast<double>(subdivision) * position +
                                                (position >= 0.0 ? 0.5 : -0.5))) /
           static_cast<double>(subdivision),
-      amount);
+      static_cast<double>(amount));
 }
 
 }  // namespace barely
